refactor: move keys into module
